+++ packages/react-magma-dom/src/components/Input/Input.stories.tsx
@@ -43,6 +43,8 @@ export const Default = Template.bind({});
 Default.args = {
   disabled: false,
   helperMessage: 'Helper message',
+  isClearable: true,
+  isInverse: false,
   placeholder: 'Placeholder text...',
 };
 Default.parameters = { controls: { exclude: ['iconPosition'] } };
@@ -56,6 +58,8 @@ Error.parameters = { controls: { exclude: ['iconPosition'] } };
 export const Large = Template.bind({});
 Large.args = {
   inputSize: InputSize.large,
+  isClearable: true,
+  isInverse: false,
 };
 Large.parameters = { controls: { exclude: ['iconPosition'] } };
 
@@ -109,14 +113,17 @@ export const ClickableIcon = Template.bind({});
 ClickableIcon.args = {
   ...Default.args,
   icon: <NotificationsIcon />,
+  iconAriaLabel: 'Notifications',
   onIconClick: () => {},
 };
+ClickableIcon.parameters = { controls: { exclude: ['iconPosition'] } };
 
 export const ClickableIconLarge = Template.bind({});
 ClickableIconLarge.args = {
   ...ClickableIcon.args,
   inputSize: InputSize.large,
 };
+ClickableIconLarge.parameters = { controls: { exclude: ['iconPosition'] } };
 
 export const Inverse = Template.bind({});
 Inverse.args = {