+++ packages/react-magma-dom/src/components/NativeSelect/NativeSelect.tsx
@@ -1,17 +1,18 @@
 import * as React from 'react';
 import styled from '../../theme/styled';
 
-import { inputBaseStyles } from '../InputBase';
+import { inputBaseStyles, inputWrapperStyles } from '../InputBase';
 import {
   FormFieldContainer,
   FormFieldContainerBaseProps,
 } from '../FormFieldContainer';
-import { SelectTriggerButton } from '../Select/SelectTriggerButton';
 
+import { DefaultDropdownIndicator } from '../Select/components';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { useIsInverse } from '../../inverse';
 import { useGenerateId } from '../../utils';
 import { ThemeInterface } from '../../theme/magma';
+import { transparentize } from 'polished';
 
 /**
  * @children required
@@ -22,14 +23,51 @@ export interface NativeSelectProps
   /**
    * @internal
    */
-  testId?: string;
   optionLabel?: string;
+  testId?: string;
 }
-const StyledNativeSelect = styled.select<{
-  theme: ThemeInterface;
+const StyledNativeSelectWrapper = styled.div<{
+  disabled?: boolean;
+  hasError?: boolean;
   isInverse?: boolean;
+  theme: ThemeInterface;
+}>`
+  ${inputWrapperStyles}
+  border:0;
+  svg {
+    color: ${props =>
+      props.isInverse && props.disabled
+        ? transparentize(0.6, props.theme.colors.neutral100)
+        : props.disabled
+        ? transparentize(0.4, props.theme.colors.neutral500)
+        : 'inherit'};
+    margin: 0 0 0 -${props => props.theme.spaceScale.spacing08};
+    pointer-events: none;
+    z-index: 1;
+  }
+`;
+
+function borderColors(props) {
+  if (props.isInverse) {
+    if (props.hasError) {
+      return props.theme.colors.danger200;
+    }
+    return transparentize(0.5, props.theme.colors.neutral100);
+  }
+  if (props.hasError) {
+    return props.theme.colors.danger;
+  }
+  return props.theme.colors.neutral500;
+}
+
+const StyledNativeSelect = styled.select<{
+  hasError?: boolean;
+  isInverse?: boolean;
+  theme: ThemeInterface;
 }>`
   ${inputBaseStyles};
+  border: 1px solid ${borderColors};
+  // Required for Windows && Chrome support
   background: inherit;
   > option {
     background: ${props =>
@@ -53,8 +91,11 @@ export const NativeSelect = React.forwardRef<HTMLDivElement, NativeSelectProps>(
       testId,
       ...other
     } = props;
+
     const theme = React.useContext(ThemeContext);
+
     const isInverse = useIsInverse(isInverseProp);
+
     const id = useGenerateId(defaultId);
 
     return (
@@ -69,23 +110,25 @@ export const NativeSelect = React.forwardRef<HTMLDivElement, NativeSelectProps>(
         messageStyle={messageStyle}
         ref={ref}
       >
-        <SelectTriggerButton
+        <StyledNativeSelectWrapper
           disabled={disabled}
           hasError={!!errorMessage}
           isInverse={isInverse}
-          toggleButtonProps={''}
+          theme={theme}
         >
           <StyledNativeSelect
-            {...other}
             data-testid={testId}
+            hasError={!!errorMessage}
             disabled={disabled}
             id={id}
             isInverse={isInverse}
             theme={theme}
+            {...other}
           >
             {children}
           </StyledNativeSelect>
-        </SelectTriggerButton>
+          <DefaultDropdownIndicator disabled={disabled} />
+        </StyledNativeSelectWrapper>
       </FormFieldContainer>
     );
   }