+++ packages/react-magma-dom/src/components/Search/Search.stories.tsx
@@ -39,6 +39,12 @@ export default {
         type: 'boolean',
       },
     },
+    isClearable: {
+      defaultValue: false,
+      control: {
+        type: 'boolean',
+      },
+    },
     isInverse: {
       defaultValue: false,
       control: {