+++ patterns/header/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 3.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [fc8a20bf]
+  - react-magma-dom@3.1.1-next.0
+
 ## 3.0.1
 
 ### Patch Changes